requirements.txt
asgiref==3.8.1
Django==4.2.3
fastapi==0.95.1
uvicorn==0.22.0
psycopg2-binary==2.9.6
gunicorn==20.1.0
google_api_python_client==2.154.0
google_auth_oauthlib==1.2.1
passlib==1.7.4
protobuf==5.29.2
pydantic==1.10.11
PyJWT==2.8.0
pytest==8.3.4
python-dotenv==1.0.1
Requests==2.32.3
SQLAlchemy==2.0.36
starlette==0.26.1
django-tailwind==3.0.0
Pillow==10.0.1
